AK+LibRegex: Move COWVector to AK
This is about to gain a new user, so move it to AK.
